refactor(frontend): improve auth security and code quality
Immediate frontend auth fixes that don't require backend changes.
Backend security infrastructure tracked in issue #165.

**Import Path Compliance**
- Fix auth components to use index file imports per CLAUDE.md
- Change '../contexts/AuthContext' → '@/contexts'
- Change '../services' → '@/services'
- Files: LoginModal, LoginForm, OAuthCallback, useOAuthPopup

**Type Safety**
- Replace 'any' with AuthUser type in useOAuthPopup hook
- Import AuthUser from '@/types/auth'

**Security Improvements**
- Remove hardcoded demo credentials from LoginForm
- Empty default email/password fields
- Remove UI text exposing demo account

**UX Enhancement**
- Add 5-minute timeout to OAuth popup flow
- Proper cleanup of timeout/interval on complete
- Clear timeout/interval before calling callbacks

Related: #165
